+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added GP Object Material Output node.
 - Added Replicate GP Layer node.
 - Added Offset GP Layer Frames node.
+- Added Texture Input node.
 
 ### Fixed